+++ toys-core/lib/toys/middleware/set_default_descriptions.rb
@@ -41,13 +41,13 @@ module Toys
       # The default description for tools.
       # @return [String]
       #
-      DEFAULT_TOOL_DESC = "(No description available)".freeze
+      DEFAULT_TOOL_DESC = "(No tool description available)".freeze
 
       ##
       # The default description for groups.
       # @return [String]
       #
-      DEFAULT_GROUP_DESC = "(A group of commands)".freeze
+      DEFAULT_GROUP_DESC = "(A group of tools)".freeze
 
       ##
       # The default description for the root tool.
@@ -68,35 +68,101 @@ module Toys
         )
 
       ##
+      # The default description for flags.
+      # @return [String]
+      #
+      DEFAULT_FLAG_DESC = "(No flag description available)".freeze
+
+      ##
+      # The default description for required args.
+      # @return [String]
+      #
+      DEFAULT_REQUIRED_ARG_DESC = "(Required argument - no description available)".freeze
+
+      ##
+      # The default description for optional args.
+      # @return [String]
+      #
+      DEFAULT_OPTIONAL_ARG_DESC = "(Optional argument - no description available)".freeze
+
+      ##
+      # The default description for remaining args.
+      # @return [String]
+      #
+      DEFAULT_REMAINING_ARG_DESC = "(Remaining arguments - no description available)".freeze
+
+      ##
       # Create a SetDefaultDescriptions middleware given default descriptions.
       #
-      # @param [String] default_tool_desc The default short description for
-      #     tools with an eecutor. Defaults to {DEFAULT_TOOL_DESC}.
+      # @param [String,nil] default_tool_desc The default short description for
+      #     tools with an executor, or `nil` not to set one. Defaults to
+      #     {DEFAULT_TOOL_DESC}.
       # @param [String,nil] default_tool_long_desc The default long description
-      #     for tools with an eecutor. If `nil` (the default), falls back to
-      #     the value of the `default_tool_desc` parameter.
-      # @param [String] default_group_desc The default short description for
-      #     tools with no eecutor. Defaults to {DEFAULT_GROUP_DESC}.
+      #     for tools with an executor, or `nil` not to set one. Defaults to
+      #     `nil`.
+      # @param [String,nil] default_group_desc The default short description
+      #     for tools with no executor, or `nil` not to set one. Defaults to
+      #     {DEFAULT_TOOL_DESC}.
       # @param [String,nil] default_group_long_desc The default long
-      #     description for tools with no eecutor. If `nil` (the default),
-      #     falls back to the value of the `default_group_desc` parameter.
-      # @param [String] default_root_desc The default short description for the
-      #     root tool. Defaults to {DEFAULT_ROOT_DESC}.
-      # @param [String] default_root_long_desc The default long description for
-      #     the root tool. Defaults to {DEFAULT_ROOT_LONG_DESC}.
+      #     description for tools with no executor, or `nil` not to set one.
+      #     Defaults to `nil`.
+      # @param [String,nil] default_root_desc The default short description for
+      #     the root tool, or `nil` not to set one. Defaults to
+      #     {DEFAULT_ROOT_DESC}.
+      # @param [String,nil] default_root_long_desc The default long description
+      #     for the root tool, or `nil` not to set one. Defaults to
+      #     {DEFAULT_ROOT_LONG_DESC}.
+      # @param [String,nil] default_flag_desc The default short description for
+      #     flags, or `nil` not to set one. Defaults to {DEFAULT_FLAG_DESC}.
+      # @param [String,nil] default_flag_long_desc The default long description
+      #     for flags, or `nil` not to set one. Defaults to `nil`.
+      # @param [String,nil] default_required_arg_desc The default short
+      #     description for required args, or `nil` not to set one. Defaults to
+      #     {DEFAULT_REQUIRED_ARG_DESC}.
+      # @param [String,nil] default_required_arg_long_desc The default long
+      #     description for required args, or `nil` not to set one. Defaults to
+      #     `nil`.
+      # @param [String,nil] default_optional_arg_desc The default short
+      #     description for optional args, or `nil` not to set one. Defaults to
+      #     {DEFAULT_OPTIONAL_ARG_DESC}.
+      # @param [String,nil] default_optional_arg_long_desc The default long
+      #     description for optional args, or `nil` not to set one. Defaults to
+      #     `nil`.
+      # @param [String,nil] default_remaining_arg_desc The default short
+      #     description for remaining args, or `nil` not to set one. Defaults
+      #     to {DEFAULT_REMAINING_ARG_DESC}.
+      # @param [String,nil] default_remaining_arg_long_desc The default long
+      #     description for remaining args, or `nil` not to set one. Defaults
+      #     to `nil`.
       #
       def initialize(default_tool_desc: DEFAULT_TOOL_DESC,
                      default_tool_long_desc: nil,
                      default_group_desc: DEFAULT_GROUP_DESC,
                      default_group_long_desc: nil,
                      default_root_desc: DEFAULT_ROOT_DESC,
-                     default_root_long_desc: DEFAULT_ROOT_LONG_DESC)
+                     default_root_long_desc: DEFAULT_ROOT_LONG_DESC,
+                     default_flag_desc: DEFAULT_FLAG_DESC,
+                     default_flag_long_desc: nil,
+                     default_required_arg_desc: DEFAULT_REQUIRED_ARG_DESC,
+                     default_required_arg_long_desc: nil,
+                     default_optional_arg_desc: DEFAULT_OPTIONAL_ARG_DESC,
+                     default_optional_arg_long_desc: nil,
+                     default_remaining_arg_desc: DEFAULT_REMAINING_ARG_DESC,
+                     default_remaining_arg_long_desc: nil)
         @default_tool_desc = default_tool_desc
         @default_tool_long_desc = default_tool_long_desc
         @default_group_desc = default_group_desc
         @default_group_long_desc = default_group_long_desc
         @default_root_desc = default_root_desc
         @default_root_long_desc = default_root_long_desc
+        @default_flag_desc = default_flag_desc
+        @default_flag_long_desc = default_flag_long_desc
+        @default_required_arg_desc = default_required_arg_desc
+        @default_required_arg_long_desc = default_required_arg_long_desc
+        @default_optional_arg_desc = default_optional_arg_desc
+        @default_optional_arg_long_desc = default_optional_arg_long_desc
+        @default_remaining_arg_desc = default_remaining_arg_desc
+        @default_remaining_arg_long_desc = default_remaining_arg_long_desc
       end
 
       ##
@@ -104,41 +170,40 @@ module Toys
       #
       def config(tool)
         if tool.root?
-          config_root_desc(tool)
+          config_descs(tool, @default_root_desc, @default_root_long_desc)
         elsif tool.includes_executor?
-          config_tool_desc(tool)
+          config_descs(tool, @default_tool_desc, @default_tool_long_desc)
         else
-          config_group_desc(tool)
+          config_descs(tool, @default_group_desc, @default_group_long_desc)
         end
+        tool.flag_definitions.each do |flag|
+          config_descs(flag, @default_flag_desc, @default_flag_long_desc)
+        end
+        config_args(tool)
         yield
       end
 
       private
 
-      def config_root_desc(tool)
-        if @default_root_long_desc && tool.long_desc.empty?
-          tool.long_desc = @default_root_long_desc
+      def config_args(tool)
+        tool.required_arg_definitions.each do |arg|
+          config_descs(arg, @default_required_arg_desc, @default_required_arg_long_desc)
         end
-        if @default_root_desc && tool.desc.empty?
-          tool.desc = @default_root_desc
+        tool.optional_arg_definitions.each do |arg|
+          config_descs(arg, @default_optional_arg_desc, @default_optional_arg_long_desc)
         end
-      end
-
-      def config_tool_desc(tool)
-        if @default_tool_long_desc && tool.long_desc.empty?
-          tool.long_desc = @default_tool_long_desc
-        end
-        if @default_tool_desc && tool.desc.empty?
-          tool.desc = @default_tool_desc
+        if tool.remaining_args_definition
+          config_descs(tool.remaining_args_definition,
+                       @default_remaining_arg_desc, @default_remaining_arg_long_desc)
         end
       end
 
-      def config_group_desc(tool)
-        if @default_group_long_desc && tool.long_desc.empty?
-          tool.long_desc = @default_group_long_desc
+      def config_descs(object, default_desc, default_long_desc)
+        if default_desc && object.desc.empty?
+          object.desc = default_desc
         end
-        if @default_group_desc && tool.desc.empty?
-          tool.desc = @default_group_desc
+        if default_long_desc && object.long_desc.empty?
+          object.long_desc = default_long_desc
         end
       end
     end